Ensure anaconda_mode can import jsonrpc, jedi
If ``sys.path`` does not include an empty string, Python won't look for modules in the same directory as ``anaconda_mode.py``. It isn't terribly uncommon for build tools/development environments to specify ``sys.path`` in a way that prevents such imports; for compatibility-sake, just make sure the empty string is present.