Fallback-behavior on dependencies installed via pip
Jedi is not frequently released, and sometimes you wish to use a more
recent version. If anaconda installs jedi at some point, it will use
that version even if there is a newer version in sys.path.

With this fix, the dependencies installed by anaconda will be used as a
last resort.