Force system scheme for installation of pip packages
According to pip documentation, as seen on some versions of Ubuntu, if
you're running installation outside of a virtual environment, the
default is to use user scheme, as if --user was specified. This,
however, conflicts with -t option.

This change forces using the system scheme to enable using install base
option, which would really just override the locations specified by user
scheme anyway. As a result, the default server installation command
becomes more robust, working as intended under wider set of
circumstances.