Check pip version during server installation
It turns out that the --system option existed only up to and including
pip 1.5.6. The behaviour changed with update to pip 6, which means:

- On pip 6 or newer, user scheme is not a default, thus not a problem.
- However, --system option was consequently removed.
- Yet, it is still needed on older pip versions.

So the default install command will now check the pip major version, and
only add --system command if it is lower than 6 (note: there are no
major versions 2 through 5 in the pip release history).