Replace usage of pip with easy_install
Pip is a wonderful utility, yet hopelessly not friendly to scripts. Its
deficiencies could be easily described as bugs, and then we could insist
on upstream fixing them. Until then, installation procedure would remain
broken in a number of configurations. Afterwards, users would need to
pay attention to use sufficiently new PIP.

The problems are at least the following:

- If PIP is installed with prefix option set in distutils.cfg, e.g., the
  way homebrew does it on mac, command line option -t will not override
  it, but instead a configuration conflict will be reported.

- Debian-based distributions patch PIP to default to using --user
  installation scheme, if not run within a virtualenv, and provide
  debian-specific --system option to override this behaviour. The
  default, yet again, conflicts with -t option, instead of being
  overridden by the latter.

By comparison, easy_install does not suffer from either of those.

This commit replaces installation using pip with easy_install. As a
result, the packages are installed as eggs, which means that server
setup launch procedure had to be replaced with a python script as well.

Additionally, since eggs are not accessible through their parent
directory and need to be added to sys.path separately, the server
detection command also needed to be altered, so that instead of just
checking the installation directory, it would check immediate
subdirectories. anaconda_mode present in one of the subdirectories would
not guarantee that it works, so it then also tries to find that path in
sys.path.

This change is entirely backward compatible:

- The check-installation-command will correctly detect legacy
  installation made by pip.
- the server-command will successfully run such installation.

Hence the the needed version of anaconda_mode python package has not yet
been modified, to allow for separate deployment and easier testing.

It is also worth noting that site.addsitedir() ends up adding eggs
before system site packages, so that the locally installed version of
python package will take precedence over any global installation. The
same applies to legacy installation in pip-style: while the absolute
path version of current directory will be added to the end of sys.path,
the current directory pattern remains in front of system directory, so
that anaconda_mode will be imported from the current directory.

This should fix #113 and #114.