Add main function to anaconda_mode.py
The emacs package will try to import this function and run it with
command line arguments, so that it won't need to know how to start the
service_factory in the future.

Currently, it does know that to allow full backward compatibility. The
python script remains backward compatible as well, hence I didn't change
version numbers.