Prevent conflicts with stdlib modules by changing working directory (#392)
Change the current directory when starting the anaconda server.

The current directory will be added to sys.path and we may end up with
files that conflict with stdlib modules. That may prevent the anaconda
server from starting:

In that case python tried to import the stdlib logging module from
foo/logging.py, which is of course bound to fail.